feat(presetData): show presetData as editable suggestions vs bunch of charts
