+++ src/Section/CompareSkills.scss
@@ -1,0 +1,5 @@
+.compare-skills__preset {
+  display: block;
+  color: #0074d9;
+  cursor: pointer;
+}
